tui: Move ui_bridge module to tui/ namespace.
